Refactor integration UI to improve organization
- Removed redundant Resources page and merged with main dashboard
- Combined Overview page with Integration Detail page
- Streamlined sidebar navigation for cleaner UX
- Reorganized integration dashboard with clearer visual hierarchy
- Fixed duplicate titles and improved spacing and layout
- Simplified UI to focus on primary actions
