+++ frontend/src/components/integration/TeamChannelSelector.tsx
@@ -274,7 +274,7 @@ const TeamChannelSelector: React.FC<TeamChannelSelectorProps> = ({
   return (
     <Box>
       <Flex mb={3} justifyContent="space-between" alignItems="center">
-        <Heading size="md">Channel Manager</Heading>
+        <Heading size="md">Channel Selection</Heading>
 
         <Button
           colorScheme="blue"
